Configure CI with Artifactory
Configure CircleCI build to automatically generate auth tokens to access
Artifactory. The `ctm.yaml` file will dictate the type of token file
that is needed and where it should be placed on the build filesystem.

The circle init script looks for the ctm.yaml file and does the token
generation.