Explained an assumption (#15)
"Ensure the keystore is loaded up with keys and certificates of the server" 

I have no idea how to do that. To be fair I didn't know how to make a keystore either, but I could google that pretty easily. But when CAS's own documentation doesn't explain it self there isn't much one can do. Took my an afternoon before I figured out what I was supposed to do to get CAS to recognize the keystore's location. I was coming to this from CAS 3 where that is managed by tomcat, so I was really confused. 

Figured it might help others to have that spelled out as well.